Allow context menus to have visible spacers
