+++ osu.Game/Graphics/UserInterface/OsuMenu.cs
@@ -6,13 +6,15 @@
 using osu.Framework.Allocation;
 using osu.Framework.Audio;
 using osu.Framework.Audio.Sample;
-using osuTK.Graphics;
 using osu.Framework.Extensions.Color4Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.UserInterface;
+using osu.Framework.Input.Events;
 using osu.Game.Graphics.Containers;
 using osuTK;
+using osuTK.Graphics;
 
 namespace osu.Game.Graphics.UserInterface
 {
@@ -78,6 +80,9 @@ namespace osu.Game.Graphics.UserInterface
             {
                 case StatefulMenuItem stateful:
                     return new DrawableStatefulMenuItem(stateful);
+
+                case OsuMenuItemSpacer spacer:
+                    return new DrawableSpacer(spacer);
             }
 
             return new DrawableOsuMenuItem(item);
@@ -89,5 +94,28 @@ namespace osu.Game.Graphics.UserInterface
         {
             Anchor = Direction == Direction.Horizontal ? Anchor.BottomLeft : Anchor.TopRight
         };
+
+        protected partial class DrawableSpacer : DrawableOsuMenuItem
+        {
+            public DrawableSpacer(MenuItem item)
+                : base(item)
+            {
+                Scale = new Vector2(1, 0.6f);
+
+                AddInternal(new Box
+                {
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    Colour = BackgroundColourHover,
+                    RelativeSizeAxes = Axes.X,
+                    Height = 2f,
+                    Width = 0.8f,
+                });
+            }
+
+            protected override bool OnHover(HoverEvent e) => true;
+
+            protected override bool OnClick(ClickEvent e) => true;
+        }
     }
 }
