+++ osu.Game/Graphics/UserInterface/OsuMenuItemSpacer.cs
@@ -1,0 +1,13 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+namespace osu.Game.Graphics.UserInterface
+{
+    public class OsuMenuItemSpacer : OsuMenuItem
+    {
+        public OsuMenuItemSpacer()
+            : base(" ")
+        {
+        }
+    }
+}
