+++ osu.Game/Screens/Edit/Components/Menus/EditorMenuBar.cs
@@ -4,11 +4,9 @@
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Events;
 using osu.Game.Graphics;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Overlays;
@@ -151,7 +149,7 @@ namespace osu.Game.Screens.Edit.Components.Menus
             {
                 switch (item)
                 {
-                    case EditorMenuItemSpacer spacer:
+                    case OsuMenuItemSpacer spacer:
                         return new DrawableSpacer(spacer);
 
                     case StatefulMenuItem stateful:
@@ -194,29 +192,6 @@ namespace osu.Game.Screens.Edit.Components.Menus
 
                     Foreground.Padding = new MarginPadding { Vertical = 2 };
                 }
-            }
-
-            private partial class DrawableSpacer : DrawableOsuMenuItem
-            {
-                public DrawableSpacer(MenuItem item)
-                    : base(item)
-                {
-                    Scale = new Vector2(1, 0.6f);
-
-                    AddInternal(new Box
-                    {
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Colour = BackgroundColourHover,
-                        RelativeSizeAxes = Axes.X,
-                        Height = 2f,
-                        Width = 0.8f,
-                    });
-                }
-
-                protected override bool OnHover(HoverEvent e) => true;
-
-                protected override bool OnClick(ClickEvent e) => true;
             }
         }
     }
