--- osu.Game/Screens/Edit/Components/Menus/EditorMenuItemSpacer.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-namespace osu.Game.Screens.Edit.Components.Menus
-{
-    public class EditorMenuItemSpacer : EditorMenuItem
-    {
-        public EditorMenuItemSpacer()
-            : base(" ")
-        {
-        }
-    }
-}
